requirements-detection.txt
# Optional: enhanced detection (YARA, ML, PE analysis)
# pip install -r requirements-detection.txt
yara-python>=4.3.0
pefile>=2023.2.7
scikit-learn>=1.3.0
PyYAML>=6.0
# Optional Windows behavioral (ETW/WMI)
pywin32>=306; sys_platform == "win32"
